Refactoring project structure to mimic the layered architecture (see description)
- Introduce "Service" package, this package will be used to contain business logics that implemented in Servlets
- Refactored Servlet codes, moving business logics to appropriate Service class (currently only applied to CustomerService and its relevant servlet)
- Moved getAllCustomer() method from SalesDAO to CustomerDAO due to earlier confusion
